+++ CHANGELOG.md
@@ -5,12 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+- Changeover to MIT license ([#2](https://github.com/scm-manager/scm-cas-plugin/pull/2))
+
 ## 2.0.0-rc5 - 2020-03-19
-### Fixes
+### Fixed
 - Redirection with anonymous access enabled fixed
 
 ## 2.0.0-rc4 - 2020-03-19
-### Fixes
+### Fixed
 - Hotfix for rc2: Redirection with anonymous access enabled fixed
 
 ## 2.0.0-rc3 - 2020-03-13
@@ -18,5 +22,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add swagger rest annotations to generate openAPI specs for the scm-openapi-plugin. ([#1](https://github.com/scm-manager/scm-cas-plugin/pull/1))
 
 ## 2.0.0-rc2 - 2020-01-31
-### Fixes
+### Fixed
 - Fix NPE on restart