+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Fixed
+- Anonymous mode after token invalidation ([#9](ihttps://github.com/scm-manager/scm-cas-plugin/pull/9))
+
 ## 2.1.1 - 2020-09-08
 ### Fixed
 - Invalidate access token cookie on cas redirect ([#8](https://github.com/scm-manager/scm-cas-plugin/pull/8))