+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 2.2.2 - 2021-02-11
 ### Fixed
 - Invalid email addresses will not lead to authentication failure ([#13](https://github.com/scm-manager/scm-cas-plugin/pull/13))